# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gym CRM API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reminders/due": {
      "get": {
        "tags": [
          "reminders"
        ],
        "summary": "Due",
        "description": "Reminders due on `day` (default: today in the gym's time zone).",
        "operationId": "due_api_reminders_due_get",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DueList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reminders/whatsapp": {
      "post": {
        "tags": [
          "reminders"
        ],
        "summary": "Log Whatsapp",
        "operationId": "log_whatsapp_api_reminders_whatsapp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatsAppLogIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reminders/history": {
      "get": {
        "tags": [
          "reminders"
        ],
        "summary": "History",
        "operationId": "history_api_reminders_history_get",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Channel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "name": "member_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Member Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_HistoryItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reminders/settings": {
      "get": {
        "tags": [
          "reminders"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_reminders_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderSettingsOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "reminders"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_reminders_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reminders/preview": {
      "post": {
        "tags": [
          "reminders"
        ],
        "summary": "Preview",
        "description": "Renders a template with sample member data, for the template editor.",
        "operationId": "preview_api_reminders_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Channel": {
        "type": "string",
        "enum": [
          "email",
          "whatsapp"
        ],
        "title": "Channel"
      },
...
      "DueList": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DueReminderOut"
            },
            "type": "array",
            "title": "Items"
          },
          "email_provider": {
            "type": "string",
            "title": "Email Provider"
          }
        },
        "type": "object",
        "required": [
          "day",
          "items",
          "email_provider"
        ],
        "title": "DueList"
      },
      "DueReminderOut": {
        "properties": {
          "membership_id": {
            "type": "string",
            "format": "uuid",
            "title": "Membership Id"
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "member_name": {
            "type": "string",
            "title": "Member Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "plan_name": {
            "type": "string",
            "title": "Plan Name"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "offset_days": {
            "type": "integer",
            "title": "Offset Days"
          },
          "kind": {
            "$ref": "#/components/schemas/Kind"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "email_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SendStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "email_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Error"
          },
          "whatsapp_sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Sent At"
          }
        },
        "type": "object",
        "required": [
          "membership_id",
          "member_id",
          "member_name",
          "phone",
          "email",
          "plan_name",
          "end_date",
          "offset_days",
          "kind",
          "subject",
          "body",
          "email_status",
          "email_error",
          "whatsapp_sent_at"
        ],
        "title": "DueReminderOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "member_name": {
            "type": "string",
            "title": "Member Name"
          },
          "plan_name": {
            "type": "string",
            "title": "Plan Name"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "offset_days": {
            "type": "integer",
            "title": "Offset Days"
          },
          "channel": {
            "$ref": "#/components/schemas/Channel"
          },
          "status": {
            "$ref": "#/components/schemas/SendStatus"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "sent_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent By Name"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "member_id",
          "member_name",
          "plan_name",
          "end_date",
          "offset_days",
          "channel",
          "status",
          "recipient",
          "error",
          "sent_by_name",
          "sent_at"
        ],
        "title": "HistoryItem"
      },
...
      "Kind": {
        "type": "string",
        "enum": [
          "before",
          "on_day",
          "after"
        ],
        "title": "Kind"
      },
...
      "Page_HistoryItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[HistoryItem]"
      },
...
      "PreviewIn": {
        "properties": {
          "template": {
            "$ref": "#/components/schemas/Template"
          },
          "kind": {
            "$ref": "#/components/schemas/Kind"
          }
        },
        "type": "object",
        "required": [
          "template",
          "kind"
        ],
        "title": "PreviewIn"
      },
      "PreviewOut": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "body"
        ],
        "title": "PreviewOut"
      },
...
      "ReminderSettingsOut": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "hour": {
            "type": "integer",
            "title": "Hour"
          },
          "offsets": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Offsets"
          },
          "templates": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Template"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Kind"
            },
            "type": "object",
            "title": "Templates"
          },
          "customized": {
            "items": {
              "$ref": "#/components/schemas/Kind"
            },
            "type": "array",
            "title": "Customized"
          },
          "placeholders": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Placeholders"
          },
          "email_provider": {
            "type": "string",
            "title": "Email Provider"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "hour",
          "offsets",
          "templates",
          "customized",
          "placeholders",
          "email_provider"
        ],
        "title": "ReminderSettingsOut"
      },
      "ReminderSettingsUpdate": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "hour": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 23.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Hour"
          },
          "offsets": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offsets"
          },
          "templates": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Template"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "propertyNames": {
                  "$ref": "#/components/schemas/Kind"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templates"
          }
        },
        "type": "object",
        "title": "ReminderSettingsUpdate"
      },
...
      "SendStatus": {
        "type": "string",
        "enum": [
          "pending",
          "sent",
          "failed"
        ],
        "title": "SendStatus"
      },
...
      "Template": {
        "properties": {
          "subject": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "body"
        ],
        "title": "Template"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WhatsAppLogIn": {
        "properties": {
          "membership_id": {
            "type": "string",
            "format": "uuid",
            "title": "Membership Id"
          },
          "offset_days": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": -60.0,
            "title": "Offset Days"
          }
        },
        "type": "object",
        "required": [
          "membership_id",
          "offset_days"
        ],
        "title": "WhatsAppLogIn"
      },
...
    }
  }
}